+++ VodopriborUchet/EditOwners.cs
@@ -52,30 +52,39 @@ namespace VodopriborUchet
         {
             if (ownersDataGridView.SelectedCells.Count > 0)
             {
-                int i = (int)ownersDataGridView.CurrentRow.Cells[0].Value;
-                using (var context = new db_sqlceEntities())
+                DialogResult dialogResult = MessageBox.Show("Вы действительно хотите удалить выбранный пункт?",
+                    "Подтвердите удаление!", MessageBoxButtons.YesNo);
+                if (dialogResult == DialogResult.Yes)
                 {
-                    try
+                    int i = (int) ownersDataGridView.CurrentRow.Cells[0].Value;
+                    using (var context = new db_sqlceEntities())
                     {
-                        var obPlc = context.owners.FirstOrDefault(p => p.id == i);
-                        if (obPlc != null)
+                        try
                         {
-                            context.owners.Remove(obPlc);
-                            context.SaveChanges();
-                            LoadData();
-                            //loadTree();
+                            var obPlc = context.owners.FirstOrDefault(p => p.id == i);
+                            if (obPlc != null)
+                            {
+                                context.owners.Remove(obPlc);
+                                context.SaveChanges();
+                                LoadData();
+                                //loadTree();
+                            }
+
+
+                        }
+                        catch (EntitySqlException ex)
+                        {
+
+                            MessageBox.Show(ex.Message);
                         }
 
-
-                    }
-                    catch (EntitySqlException ex)
-                    {
-
-                        MessageBox.Show(ex.Message);
                     }
 
                 }
-
+                else if (dialogResult == DialogResult.No)
+                {
+                    return;
+                }
             }
         }
